Poor man's typing of Topic.extras[siteId]
Not ideal but better than previous solution, which was basically Record<str, any>. This way we still get type hints when accessing `topic.value.extras[extrasToProcess].theme` for example.

This is a hack using explicit casting to `SiteId` of `extrasToProcess` from config file, with a dummy litteral type of `siteId`. The only value passed around is this one, so it matches with the type definition.